finished demo app

+++ src/main/java/com/mvpmatch/vendingmachine/controller/ProductController.java
@@ -1,17 +1,24 @@
 package com.mvpmatch.vendingmachine.controller;
 
+import com.mvpmatch.vendingmachine.config.model.MyUserPrincipal;
 import com.mvpmatch.vendingmachine.entity.ProductEntity;
+import com.mvpmatch.vendingmachine.entity.Role;
 import com.mvpmatch.vendingmachine.service.ProductService;
 import java.util.Map;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+
+import static com.mvpmatch.vendingmachine.util.Util.checkIfRoleIsValid;
 
 /**
  * Created by vedadsurkovic on 3/28/22
  **/
 @RestController
 @RequestMapping("/product")
+@PreAuthorize("isAuthenticated()")
 public class ProductController {
 
     private final ProductService productService;
@@ -21,8 +28,13 @@ public class ProductController {
     }
 
     @PostMapping
-    @PreAuthorize("hasAuthority('SELLER')")
     public ResponseEntity<?> create(@RequestBody final ProductEntity productEntity) {
+        final Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        final MyUserPrincipal myUserPrincipal = (MyUserPrincipal) auth.getPrincipal();
+
+        if (checkIfRoleIsValid(myUserPrincipal, Role.SELLER))
+            return ResponseEntity.badRequest().body("Only user with role SELLER can access this endpoint");
+
         return ResponseEntity.ok(productService.create(productEntity));
     }
 
@@ -37,23 +49,38 @@ public class ProductController {
     }
 
     @PutMapping
-    @PreAuthorize("hasAuthority('SELLER')")
     public ResponseEntity<?> update(@RequestBody final ProductEntity productEntity) {
+        final Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        final MyUserPrincipal myUserPrincipal = (MyUserPrincipal) auth.getPrincipal();
+
+        if (checkIfRoleIsValid(myUserPrincipal, Role.SELLER))
+            return ResponseEntity.badRequest().body("Only user with role SELLER can access this endpoint");
+
         return ResponseEntity.ok(productService.update(productEntity));
     }
 
     @DeleteMapping("/{productId}")
-    @PreAuthorize("hasAuthority('SELLER')")
     public ResponseEntity<?> delete(@PathVariable final Long productId) {
+        final Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        final MyUserPrincipal myUserPrincipal = (MyUserPrincipal) auth.getPrincipal();
+
+        if (checkIfRoleIsValid(myUserPrincipal, Role.SELLER))
+            return ResponseEntity.badRequest().body("Only user with role SELLER can access this endpoint");
+
         productService.delete(productId);
         return ResponseEntity.ok().build();
     }
 
     @PatchMapping("/buy")
-    @PreAuthorize("hasAuthority('BUYER')")
     public ResponseEntity<?> buy(@RequestBody final Map<String, String> request) {
+        final Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        final MyUserPrincipal myUserPrincipal = (MyUserPrincipal) auth.getPrincipal();
+
+        if (checkIfRoleIsValid(myUserPrincipal, Role.BUYER))
+            return ResponseEntity.badRequest().body("Only user with role BUYER can access this endpoint");
+
         return ResponseEntity.ok(productService.buy(
-            Long.valueOf(request.get("userId")),
+            myUserPrincipal.getUser().getId(),
             Long.valueOf(request.get("productId")),
             Integer.valueOf(request.get("amount"))));
     }
